Move `ahbicht.mapping_results` to `ahbicht.models.mapping_results`
and `parse_repeatability` from `ahbicht.mapping_results` to `ahbicht.utility_functions`
